+++ Source/GridModule.h
@@ -82,7 +82,7 @@ public:
    //IGridController
    void SetGridControllerOwner(IGridControllerListener* owner) override { mGridControllerOwner = owner; }
    void SetLight(int x, int y, GridColor color, bool force = false) override;
-   void SetLightDirect(int x, int y, int color, bool force = false) override;
+   void SetLightDirect(int x, int y, double color, bool force = false) override;
    void ResetLights() override;
    int NumCols() override { return GetCols(); }
    int NumRows() override { return GetRows(); }
