+++ Source/NoteExpressionRouter.cpp
@@ -39,7 +39,7 @@ NoteExpressionRouter::NoteExpressionRouter()
    for (auto i = 0; i < kMaxDestinations; ++i)
    {
       mExpressions[i].register_symbol_table(mSymbolTable);
-      auto p = exprtk::parser<float>();
+      auto p = exprtk::parser<double>();
       p.compile("1", mExpressions[i]);
    }
 }
@@ -115,7 +115,7 @@ void NoteExpressionRouter::TextEntryComplete(TextEntry* entry)
    {
       if (entry == mExpressionWidget[i])
       {
-         auto p = exprtk::parser<float>();
+         auto p = exprtk::parser<double>();
          if (!p.compile(entry->GetText(), mExpressions[i]))
          {
             ofLog() << "Error parsing expression '" << entry->GetText() << "' " << p.error();
